reject: optimize
