+++ frontend/src/__data__/service/account.api.ts
@@ -1,24 +1,6 @@
-import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
-import {
-  Accounts,
-  Account,
-  AccountHistory,
-  ExpandSearchAdmissionsBody,
-  ChangePasswordData,
-  RootState
-} from '../../types/api'
-import { config } from '../config'
-
-const baseQuery = fetchBaseQuery({
-  baseUrl: config.baseAPI,
-  prepareHeaders: (headers, { getState }) => {
-    const token = (getState() as RootState).auth.token
-    if (token) {
-      headers.set('Authorization', `Bearer ${token}`)
-    }
-    return headers
-  }
-})
+import { createApi } from '@reduxjs/toolkit/query/react'
+import { Accounts, Account, AccountHistory, ExpandSearchAdmissionsBody, ChangePasswordData } from '../../types/api'
+import { baseQuery } from '../utils'
 
 export const apiAccount = createApi({
   reducerPath: 'apiAccount',
