+++ frontend/src/__data__/service/object.api.ts
@@ -1,18 +1,6 @@
-import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
-import { ObjectHistory, Objects, RootState, soloObject } from '../../types/api'
-import { config } from '../config'
-
-const baseQuery = fetchBaseQuery({
-  baseUrl: config.baseAPI,
-  prepareHeaders: (headers, { getState }) => {
-    console.log((getState() as RootState).auth.token)
-    const token = (getState() as RootState).auth.token
-    if (token) {
-      headers.set('Authorization', `Bearer ${token}`)
-    }
-    return headers
-  }
-})
+import { createApi } from '@reduxjs/toolkit/query/react'
+import { ObjectHistory, Objects, soloObject } from '../../types/api'
+import { baseQuery } from '../utils'
 
 export const apiObject = createApi({
   reducerPath: 'apiObject',
