+++ frontend/src/components/smart-table/basic/index.tsx
@@ -6,8 +6,6 @@ import TableRow from '@mui/material/TableRow'
 import { ButtonNames, ShortcutButtons } from '../../shortcut-buttons'
 import { Box } from '@mui/material'
 import { Size } from '..'
-import { useContext } from 'react'
-import { objectsContext } from '../../../contexts/api'
 import { Objects, Admissions } from '../../../types/api'
 import { useGetAllObjectsQuery } from '../../../__data__/service/object.api'
 
@@ -20,20 +18,6 @@ type URL = {
 export const Basic = ({ currentURL, buttonNames, size }: URL & ButtonNames & { size: Size }) => {
   const objectsURL = currentURL === '/objects'
 
-  // const {
-  //   objectData = [
-  //     {
-  //       id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
-  //       name: 'ГМЦ'
-  //     },
-  //     {
-  //       id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
-  //       name: 'ГМЦ'
-  //     },
-  //   ],
-  //   objectError,
-  //   objectLoading
-  // } = useContext(objectsContext)
   const { data: objectData, error: objectError, isLoading: objectLoading } = useGetAllObjectsQuery()
 
   return (
