--- frontend/src/contexts/api/index.ts
@@ -1,6 +1,0 @@
-import { createContext } from 'react'
-
-export const accountsContext = createContext<any>([])
-export const objectsContext = createContext<any>([])
-export const admissionssContext = createContext<any>([])
-
